# app/organization/schemas/building.py
from datetime import datetime
from pydantic import BaseModel


class BuildingBase(BaseModel):
    address: str
    latitude: float
    longitude: float


class BuildingSchema(BuildingBase):
    id: int
    created_at: datetime
    updated_at: datetime

    model_config = {"from_attributes": True}